requirements.txt
# pycuda # if cuda is installed
six==1.10.0
cupy==1.0.2
chainer==2.0.2
cycler==0.10.0
filelock==2.0.7
#matplotlib==1.5.3
nose==1.3.7
numpy==1.11.2
Pillow==3.4.2
protobuf==3.1.0.post1
pyparsing==2.1.10
python-dateutil==2.6.0
pytz==2016.7
